Force fp32 for Seed-TTS WER Whisper eval to fix transformers 5.x dtype mismatch
Buildkite #8418 Omni Accuracy Test failed with conv1 dtype mismatch
("Input type (float) and bias type (c10::Half) should be the same")
for every WER sample. Root cause: transformers 5.x honors
`model.config.torch_dtype` in `from_pretrained` when no explicit
torch_dtype is passed, and openai/whisper-large-v3 ships
torch_dtype=float16. The Seed-TTS eval protocol (and the helper
`_transcribe_en_f32_16k`) feed fp32 input_features from
WhisperProcessor, so conv1d weights/bias must also be fp32.

Pin torch_dtype=torch.float32 explicitly in `_ensure_en_asr`.

+++ vllm_omni/benchmarks/data_modules/seed_tts_eval.py
@@ -351,6 +351,7 @@ def _ensure_en_asr() -> None:
     with _lock:
         if _en_processor is not None:
             return
+        import torch
         from transformers import WhisperForConditionalGeneration, WhisperProcessor
 
         _device = _get_eval_device()
@@ -361,7 +362,13 @@ def _ensure_en_asr() -> None:
             _device,
         )
         _en_processor = WhisperProcessor.from_pretrained(mid)
-        _en_model = WhisperForConditionalGeneration.from_pretrained(mid).to(_device)
+        # Force float32 weights/bias to match `_transcribe_en_f32_16k`'s fp32 protocol
+        # and `WhisperProcessor`'s default fp32 ``input_features``. transformers >=5.x
+        # honors ``model.config.torch_dtype`` in ``from_pretrained``, and
+        # ``openai/whisper-large-v3`` ships ``torch_dtype: float16`` in its config —
+        # without this override conv1 raises
+        # ``RuntimeError: Input type (float) and bias type (c10::Half) should be the same``.
+        _en_model = WhisperForConditionalGeneration.from_pretrained(mid, torch_dtype=torch.float32).to(_device)
         _en_model.eval()
